chore: exercicio 16 && 17 - cria o controller que deleta, e rota no app 'DELETE' /books/:id

+++ T19C-Back-End/bloco-24-NodeJs-ORM-e-Autenticacao/dia-01-ORM-Interface-da-aplicacao-com-o-banco-de-dados/src/services/BooksService.js
@@ -9,6 +9,10 @@ const create = async ({ title, author, pageQuantity }) => {
 }
 
 const destroy = async (id) => {
+  const isValid = getById(id);
+
+  if (!isValid || isValid === undefined) return null
+
   const result = await Book.destroy({ where: { id } });
 
   if (!result) return null;
